requirements.txt
# Combined requirements for Event Manager project

# Frontend Dependencies
toga-winforms>=0.5.0
requests>=2.31.0

# Backend Dependencies  
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Development Dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Production Dependencies
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
